rename file connection.py to ami_protocol.py
